Error fixing for MSVC 2012 + deails on README

+++ include/Ensemble.h
@@ -47,7 +47,7 @@ public:
 
 protected:
     ens_type type;
-    double N;
+    int N;
     double P;
     double V;
     double T;
